♻️  F09:  ADMIN > Create reusable form in to EditForm and AddForm
