+++ src/components/pages/order/Main/Basket.js
@@ -1,0 +1,13 @@
+import styled from "styled-components"
+
+function Basket() {
+  return (
+    <BasketStyled>Basket</BasketStyled>
+  )
+}
+
+export default Basket
+
+const BasketStyled = styled.div`
+  background: pink;
+`
