+++ src/components/pages/order/Main/MainRightSide/Admin/Admin.js
@@ -3,7 +3,7 @@ import styled from "styled-components";
 import AdminTabs from "./AdminTabs";
 import AdminPanel from "./AdminPanel/AdminPanel";
 import { useContext } from "react";
-import OrderContext from "../../../../../context/OrderContext";
+import OrderContext from "../../../../../../context/OrderContext";
