+++ src/components/pages/order/Main/MainRightSide/Admin/AdminPanel/EditForm.js
@@ -1,0 +1,27 @@
+import { useContext } from "react";
+import OrderContext from "../../../../../../../context/OrderContext";
+import EditInfoMessage from "./EditInfoMessage";
+import Form from "./Form";
+
+const EditForm = () => {
+    // state
+    const { productSelected, setProductSelected, handleEdit, titleEditRef } = useContext(OrderContext);
+    
+    // comportements (gestionnaires d'événement ou "event handlers")
+    const handleChange = (e) => {
+        const {name, value} = e.target;
+        const productBeingUpdated = {...productSelected, [name]: value }
+        
+        setProductSelected(productBeingUpdated) // cette ligne update le formulaire
+        handleEdit(productBeingUpdated, e) // cette ligne update le menu
+    }
+    
+    // affichage
+    return ( 
+        <Form product = {productSelected} onChange = {handleChange} ref={titleEditRef}>
+            <EditInfoMessage/>
+        </Form>
+     );
+}
+ 
+export default EditForm;
