+++ src/components/pages/order/Main/MainRightSide/Admin/AdminPanel/ImagePreview.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import styled from 'styled-components';
-import { theme } from '../../../../../../theme';
+import { theme } from '../../../../../../../theme';
 
 function ImagePreview({imageSource, title}) {
   return (
