+++ src/components/pages/order/Main/MainRightSide/Admin/AdminPanel/SubmitMessage.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components';
 import { FiCheck } from "react-icons/fi";
-import { theme } from '../../../../../../theme';
+import { theme } from '../../../../../../../theme';
 
 function SubmitMessage() {
   return (
