+++ src/components/pages/order/Main/MainRightSide/MainRightSide.js
@@ -1,0 +1,27 @@
+import { useContext } from "react"
+import styled from "styled-components"
+import { theme } from "../../../../../theme"
+import Admin from "./Admin/Admin"
+import Menu from "./Menu/Menu"
+import OrderContext from "../../../../../context/OrderContext"
+
+function MainRightSide() {
+    const {isModeAdmin} = useContext(OrderContext);
+
+  return (
+    <MainRightSideStyled>
+        <Menu/>
+        {isModeAdmin && <Admin />}
+    </MainRightSideStyled>
+  )
+}
+
+export default MainRightSide
+
+const MainRightSideStyled = styled.div`
+  position: relative;
+  overflow-y: hidden;
+  display: grid;
+  border-bottom-left-radius: ${theme.borderRadius.extraRound};
+  border-bottom-right-radius: ${theme.borderRadius.extraRound};
+`
