+++ src/components/pages/order/Main/MainRightSide/Menu/EmptyMenuAdmin.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components';
-import { theme } from '../../../../../theme';
-import Button from '../../../../reusable-ui/Button';
+import { theme } from '../../../../../../theme';
+import Button from '../../../../../reusable-ui/Button';
 
 function EmptyMenuAdmin({onReset}) {
